+++ packages/swingset-liveslots/test/test-liveslots.js
@@ -1,4 +1,3 @@
-/* global WeakRef */
 import test from 'ava';
 import '@endo/init/debug.js';
 
@@ -6,21 +5,10 @@ import { E } from '@endo/eventual-send';
 import { Far } from '@endo/marshal';
 import { makePromiseKit } from '@endo/promise-kit';
 import { Fail } from '@agoric/assert';
-import engineGC from './engine-gc.js';
-import { makeGcAndFinalize } from './gc-and-finalize.js';
 import { makeLiveSlots, makeMarshaller } from '../src/liveslots.js';
 import { kslot, kser, kunser } from './kmarshal.js';
 import { buildSyscall, makeDispatch } from './liveslots-helpers.js';
-import {
-  makeMessage,
-  makeStartVat,
-  makeBringOutYourDead,
-  makeResolve,
-  makeReject,
-  makeDropExports,
-  makeRetireExports,
-  makeRetireImports,
-} from './util.js';
+import { makeMessage, makeStartVat, makeResolve, makeReject } from './util.js';
 import { makeMockGC } from './mock-gc.js';
 
 function matchIDCounterSet(t, log) {
@@ -302,45 +290,6 @@ test('liveslots retires outbound promise IDs after resolve to data', async t => 
 
 test('liveslots retires outbound promise IDs after reject', async t => {
   await doOutboundPromise(t, 'reject');
-});
-
-test('liveslots retains pending exported promise', async t => {
-  const gcAndFinalize = makeGcAndFinalize(engineGC);
-  const { log, syscall } = buildSyscall();
-  let watch;
-  const success = [];
-  function build(_vatPowers) {
-    const root = Far('root', {
-      make() {
-        const pk = makePromiseKit();
-        watch = new WeakRef(pk.promise);
-        // we export the Promise, but do not retain resolve/reject
-        return [pk.promise];
-      },
-      // if liveslots fails to keep a strongref to the Promise, it will have
-      // been collected by now, and calling check() will fail, because
-      // liveslots can't create a new Promise import when the allocatedByVat
-      // says it was an export
-      check(_p) {
-        success.push('yes');
-      },
-    });
-    return root;
-  }
-
-  const { dispatch } = await makeDispatch(syscall, build);
-  log.length = 0; // assume pre-build vatstore operations are correct
-  const rootA = 'o+0';
-  const resultP = 'p-1';
-  await dispatch(makeMessage(rootA, 'make', [], resultP));
-  await gcAndFinalize();
-  t.truthy(watch.deref(), 'Promise not retained');
-  t.is(log[0].type, 'resolve');
-  const res0 = log[0].resolutions[0];
-  t.is(res0[0], resultP);
-  const exportedVPID = res0[2].slots[0]; // currently p+5
-  await dispatch(makeMessage(rootA, 'check', [kslot(exportedVPID)]));
-  t.deepEqual(success, ['yes']);
 });
 
 function hush(p) {
@@ -1019,327 +968,6 @@ test('disavow', async t => {
   t.deepEqual(log, []);
 });
 
-test('liveslots retains device nodes', async t => {
-  const gcAndFinalize = makeGcAndFinalize(engineGC);
-  const { syscall } = buildSyscall();
-  let watch;
-  const recognize = new WeakSet(); // real WeakSet
-  const success = [];
-  function build(_vatPowers) {
-    const root = Far('root', {
-      first(dn) {
-        watch = new WeakRef(dn);
-        recognize.add(dn);
-      },
-      second(dn) {
-        success.push(recognize.has(dn));
-      },
-    });
-    return root;
-  }
-
-  const { dispatch } = await makeDispatch(syscall, build);
-  const rootA = 'o+0';
-  const device = 'd-1';
-  await dispatch(makeMessage(rootA, 'first', [kslot(device)]));
-  await gcAndFinalize();
-  t.truthy(watch.deref(), 'Device node not retained');
-  await dispatch(makeMessage(rootA, 'second', [kslot(device)]));
-  t.deepEqual(success, [true]);
-});
-
-test('GC syscall.dropImports', async t => {
-  const { log, syscall } = buildSyscall();
-  let wr;
-  function build(_vatPowers) {
-    // eslint-disable-next-line no-unused-vars
-    let presence1;
-    const root = Far('root', {
-      one(arg) {
-        presence1 = arg;
-        wr = new WeakRef(arg);
-      },
-      two() {},
-      three() {
-        presence1 = undefined; // drops the import
-      },
-    });
-    return root;
-  }
-  const { dispatch } = await makeDispatch(syscall, build, 'vatA', {
-    enableDisavow: true,
-  });
-  log.length = 0; // assume pre-build vatstore operations are correct
-  const rootA = 'o+0';
-  const arg = 'o-1';
-
-  // tell the vat make a Presence and hold it for a moment
-  // rp1 = root~.one(arg)
-  await dispatch(makeMessage(rootA, 'one', [kslot(arg)]));
-  await dispatch(makeBringOutYourDead());
-  t.truthy(wr.deref());
-
-  // an intermediate message will trigger GC, but the presence is still held
-  await dispatch(makeMessage(rootA, 'two', []));
-  await dispatch(makeBringOutYourDead());
-  t.truthy(wr.deref());
-
-  // now tell the vat to drop the 'arg' presence we gave them earlier
-  await dispatch(makeMessage(rootA, 'three', []));
-  await dispatch(makeBringOutYourDead());
-
-  // the presence itself should be gone
-  t.falsy(wr.deref());
-
-  // first it will check that there are no VO's holding onto it
-  const l2 = log.shift();
-  t.deepEqual(l2, {
-    type: 'vatstoreGet',
-    key: 'vom.rc.o-1',
-    result: undefined,
-  });
-
-  const l3 = log.shift();
-  t.deepEqual(l3, {
-    type: 'vatstoreGetNextKey',
-    priorKey: 'vom.ir.o-1|',
-    result: 'vom.rc.o+d6/1',
-  });
-
-  // since nothing else is holding onto it, the vat should emit a dropImports
-  const l4 = log.shift();
-  t.deepEqual(l4, {
-    type: 'dropImports',
-    slots: [arg],
-  });
-
-  // and since the vat never used the Presence in a WeakMap/WeakSet, they
-  // cannot recognize it either, and will emit retireImports
-  const l5 = log.shift();
-  t.deepEqual(l5, {
-    type: 'retireImports',
-    slots: [arg],
-  });
-
-  t.deepEqual(log, []);
-});
-
-test('GC dispatch.retireImports', async t => {
-  const { log, syscall } = buildSyscall();
-  function build(_vatPowers) {
-    // eslint-disable-next-line no-unused-vars
-    let presence1;
-    const root = Far('root', {
-      one(arg) {
-        presence1 = arg;
-      },
-    });
-    return root;
-  }
-  const { dispatch } = await makeDispatch(syscall, build, 'vatA', {
-    enableDisavow: true,
-  });
-  log.length = 0; // assume pre-build vatstore operations are correct
-  const rootA = 'o+0';
-  const arg = 'o-1';
-
-  // tell the vat make a Presence and hold it
-  // rp1 = root~.one(arg)
-  await dispatch(makeMessage(rootA, 'one', [kslot(arg)]));
-
-  // when the upstream export goes away, the kernel will send a
-  // dispatch.retireImport into the vat
-  await dispatch(makeRetireImports(arg));
-  // for now, we only care that it doesn't crash
-  t.like(log.shift(), { type: 'vatstoreGetNextKey' });
-  t.deepEqual(log, []);
-
-  // when we implement VOM.vrefIsRecognizable, this test might do more
-});
-
-test('GC dispatch.retireExports', async t => {
-  const { log, syscall } = buildSyscall();
-  function build(_vatPowers) {
-    const ex1 = Far('export', {});
-    const root = Far('root', {
-      one() {
-        return ex1;
-      },
-    });
-    return root;
-  }
-  const { dispatch } = await makeDispatch(syscall, build, 'vatA', {
-    enableDisavow: true,
-  });
-  log.length = 0; // assume pre-build vatstore operations are correct
-  const rootA = 'o+0';
-
-  // rp1 = root~.one()
-  // ex1 = await rp1
-  const rp1 = 'p-1';
-  await dispatch(makeMessage(rootA, 'one', [], rp1));
-  const l1 = log.shift();
-  const ex1 = l1.resolutions[0][2].slots[0];
-  t.deepEqual(l1, {
-    type: 'resolve',
-    resolutions: [[rp1, false, kser(kslot(ex1, 'export'))]],
-  });
-  matchIDCounterSet(t, log);
-  t.deepEqual(log, []);
-
-  // All other vats drop the export, but since the vat holds it strongly, the
-  // vat says nothing
-  await dispatch(makeDropExports(ex1));
-  t.deepEqual(log, []);
-
-  // Also, all other vats cease to be able to recognize it, which will delete
-  // the clist entry and allows the vat to delete some slotToVal tables. The
-  // vat does not need to react, but we want to make sure the dispatch
-  // doesn't crash anything.
-  await dispatch(makeRetireExports(ex1));
-  t.deepEqual(log, []);
-});
-
-test('dropImports', async t => {
-  const { syscall } = buildSyscall();
-  const imports = [];
-  const gcTools = makeMockGC();
-
-  function build(_vatPowers) {
-    const root = Far('root', {
-      hold(imp) {
-        imports.push(imp);
-      },
-      free() {
-        gcTools.kill(imports.pop());
-      },
-      ignore(imp) {
-        gcTools.kill(imp);
-      },
-    });
-    return root;
-  }
-
-  const ls = makeLiveSlots(syscall, 'vatA', {}, {}, gcTools, undefined, () => ({
-    buildRootObject: build,
-  }));
-  const { dispatch, possiblyDeadSet } = ls;
-  await dispatch(makeStartVat(kser()));
-  const allFRs = gcTools.getAllFRs();
-  t.is(allFRs.length, 2);
-  const FR = allFRs[0];
-
-  const rootA = 'o+0';
-
-  // immediate drop should push import to possiblyDeadSet after finalizer runs
-  await dispatch(makeMessage(rootA, 'ignore', [kslot('o-1')]));
-  // the immediate gcTools.kill() means that the import should now be in the
-  // "COLLECTED" state
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 1);
-  FR.runOneCallback(); // moves to FINALIZED
-  t.deepEqual(possiblyDeadSet, new Set(['o-1']));
-  possiblyDeadSet.delete('o-1'); // pretend liveslots did syscall.dropImport
-
-  // separate hold and free should do the same
-  await dispatch(makeMessage(rootA, 'hold', [kslot('o-2')]));
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 0);
-  await dispatch(makeMessage(rootA, 'free', []));
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 1);
-  FR.runOneCallback(); // moves to FINALIZED
-  t.deepEqual(possiblyDeadSet, new Set(['o-2']));
-  possiblyDeadSet.delete('o-2'); // pretend liveslots did syscall.dropImport
-
-  // re-introduction during COLLECTED should return to REACHABLE
-
-  await dispatch(makeMessage(rootA, 'ignore', [kslot('o-3')]));
-  // now COLLECTED
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 1);
-
-  await dispatch(makeMessage(rootA, 'hold', [kslot('o-3')]));
-  // back to REACHABLE
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 1);
-
-  FR.runOneCallback(); // stays at REACHABLE
-  t.deepEqual(possiblyDeadSet, new Set());
-
-  await dispatch(makeMessage(rootA, 'free', []));
-  // now COLLECTED
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 1);
-
-  FR.runOneCallback(); // moves to FINALIZED
-  t.deepEqual(possiblyDeadSet, new Set(['o-3']));
-  possiblyDeadSet.delete('o-3'); // pretend liveslots did syscall.dropImport
-
-  // multiple queued finalizers are idempotent, remains REACHABLE
-
-  await dispatch(makeMessage(rootA, 'ignore', [kslot('o-4')]));
-  // now COLLECTED
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 1);
-
-  await dispatch(makeMessage(rootA, 'ignore', [kslot('o-4')]));
-  // moves to REACHABLE and then back to COLLECTED
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 2);
-
-  await dispatch(makeMessage(rootA, 'hold', [kslot('o-4')]));
-  // back to REACHABLE
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 2);
-
-  FR.runOneCallback(); // stays at REACHABLE
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 1);
-
-  FR.runOneCallback(); // stays at REACHABLE
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 0);
-
-  // multiple queued finalizers are idempotent, remains FINALIZED
-
-  await dispatch(makeMessage(rootA, 'ignore', [kslot('o-5')]));
-  // now COLLECTED
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 1);
-
-  await dispatch(makeMessage(rootA, 'ignore', [kslot('o-5')]));
-  // moves to REACHABLE and then back to COLLECTED
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 2);
-
-  FR.runOneCallback(); // moves to FINALIZED
-  t.deepEqual(possiblyDeadSet, new Set(['o-5']));
-  t.is(FR.countCallbacks(), 1);
-
-  FR.runOneCallback(); // stays at FINALIZED
-  t.deepEqual(possiblyDeadSet, new Set(['o-5']));
-  t.is(FR.countCallbacks(), 0);
-  possiblyDeadSet.delete('o-5'); // pretend liveslots did syscall.dropImport
-
-  // re-introduction during FINALIZED moves back to REACHABLE
-
-  await dispatch(makeMessage(rootA, 'ignore', [kslot('o-6')]));
-  // moves to REACHABLE and then back to COLLECTED
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 1);
-
-  FR.runOneCallback(); // moves to FINALIZED
-  t.deepEqual(possiblyDeadSet, new Set(['o-6']));
-  t.is(FR.countCallbacks(), 0);
-
-  await dispatch(makeMessage(rootA, 'hold', [kslot('o-6')]));
-  await dispatch(makeBringOutYourDead());
-  // back to REACHABLE, removed from possiblyDeadSet
-  t.deepEqual(possiblyDeadSet, new Set());
-  t.is(FR.countCallbacks(), 0);
-});
-
 test('buildVatNamespace not called until after startVat', async t => {
   const { syscall } = buildSyscall();
   const gcTools = makeMockGC();
@@ -1356,150 +984,6 @@ test('buildVatNamespace not called until after startVat', async t => {
   t.falsy(buildCalled);
   await ls.dispatch(makeStartVat(kser()));
   t.truthy(buildCalled);
-});
-
-test('GC dispatch.dropExports', async t => {
-  const { log, syscall } = buildSyscall();
-  let wr;
-  function build(_vatPowers) {
-    const root = Far('root', {
-      one() {
-        const ex1 = Far('export', {});
-        wr = new WeakRef(ex1);
-        return ex1;
-        // ex1 goes out of scope, dropping last userspace strongref
-      },
-      two() {},
-    });
-    return root;
-  }
-  const { dispatch } = await makeDispatch(syscall, build, 'vatA', {
-    enableDisavow: true,
-  });
-  log.length = 0; // assume pre-build vatstore operations are correct
-  const rootA = 'o+0';
-
-  // rp1 = root~.one()
-  // ex1 = await rp1
-  const rp1 = 'p-1';
-  await dispatch(makeMessage(rootA, 'one', [], rp1));
-  await dispatch(makeBringOutYourDead());
-  const l1 = log.shift();
-  const ex1 = l1.resolutions[0][2].slots[0];
-  t.deepEqual(l1, {
-    type: 'resolve',
-    resolutions: [[rp1, false, kser(kslot(ex1, 'export'))]],
-  });
-  t.deepEqual(log.shift(), {
-    type: 'vatstoreSet',
-    key: 'idCounters',
-    value: '{"exportID":11,"collectionID":5,"promiseID":5}',
-  });
-  t.deepEqual(log, []);
-
-  // the exported Remotable should be held in place by exportedRemotables
-  // until we tell the vat we don't need it any more
-  t.truthy(wr.deref());
-
-  // an intermediate message will trigger GC, but the presence is still held
-  await dispatch(makeMessage(rootA, 'two', []));
-  await dispatch(makeBringOutYourDead());
-  t.truthy(wr.deref());
-
-  // now tell the vat we don't need a strong reference to that export.
-  await dispatch(makeDropExports(ex1));
-  await dispatch(makeBringOutYourDead());
-
-  // that should allow ex1 to be collected
-  t.falsy(wr.deref());
-
-  // and once it's collected, the vat should emit `syscall.retireExport`
-  // because nobody else will be able to recognize it again
-  const l2 = log.shift();
-  t.deepEqual(l2, {
-    type: 'retireExports',
-    slots: [ex1],
-  });
-  t.deepEqual(log, []);
-});
-
-test('GC dispatch.retireExports inhibits syscall.retireExports', async t => {
-  const { log, syscall } = buildSyscall();
-  let wr;
-  function build(_vatPowers) {
-    let ex1;
-    const root = Far('root', {
-      hold() {
-        ex1 = Far('export', {});
-        wr = new WeakRef(ex1);
-        return ex1;
-      },
-      two() {},
-      drop() {
-        // eslint-disable-next-line no-unused-vars
-        ex1 = undefined; // drop the last userspace strongref
-      },
-    });
-    return root;
-  }
-  const { dispatch } = await makeDispatch(syscall, build, 'vatA', {
-    enableDisavow: true,
-  });
-  log.length = 0; // assume pre-build vatstore operations are correct
-  const rootA = 'o+0';
-
-  // rp1 = root~.hold()
-  // ex1 = await rp1
-  const rp1 = 'p-1';
-  await dispatch(makeMessage(rootA, 'hold', [], rp1));
-  await dispatch(makeBringOutYourDead());
-  const l1 = log.shift();
-  const ex1 = l1.resolutions[0][2].slots[0];
-  t.deepEqual(l1, {
-    type: 'resolve',
-    resolutions: [[rp1, false, kser(kslot(ex1, 'export'))]],
-  });
-  t.deepEqual(log.shift(), {
-    type: 'vatstoreSet',
-    key: 'idCounters',
-    value: '{"exportID":11,"collectionID":5,"promiseID":5}',
-  });
-  t.deepEqual(log, []);
-
-  // the exported Remotable should be held in place by exportedRemotables
-  // until we tell the vat we don't need it any more
-  t.truthy(wr.deref());
-
-  // an intermediate message will trigger GC, but the presence is still held
-  await dispatch(makeMessage(rootA, 'two', []));
-  await dispatch(makeBringOutYourDead());
-  t.truthy(wr.deref());
-
-  // now tell the vat we don't need a strong reference to that export.
-  await dispatch(makeDropExports(ex1));
-  await dispatch(makeBringOutYourDead());
-
-  // that removes the liveslots strongref, but the vat's remains in place
-  t.truthy(wr.deref());
-
-  // now the kernel tells the vat we can't even recognize the export
-  await dispatch(makeRetireExports(ex1));
-  await dispatch(makeBringOutYourDead());
-
-  // that ought to delete the table entry, but doesn't affect the vat
-  // strongref
-  t.truthy(wr.deref());
-
-  // now tell the vat to drop its strongref
-  await dispatch(makeMessage(rootA, 'drop', []));
-  await dispatch(makeBringOutYourDead());
-
-  // which should let the export be collected
-  t.falsy(wr.deref());
-
-  // the vat should *not* emit `syscall.retireExport`, because it already
-  // received a dispatch.retireExport
-  t.deepEqual(log, []);
 });
 
 // todo: test that ancillary promises cause syscall.resolve, also test
